Move composables of `difficulty selection` screen into `composable` package
