+++ ui-difficulty-selection/src/main/java/com/jayasuryat/difficultyselection/composable/StartButton.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.jayasuryat.difficultyselection
+package com.jayasuryat.difficultyselection.composable
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
